requirements.txt
tensorflow
streamlit
pillow
numpy
requests
pandas
google-cloud-bigquery
runipy

#for schedule job 
python-crontab
schedule

#Run this file with - pip3 install -r requirements.txt